DreamJob added input: image

+++ src/main/java/org/dreamjob/service/SimpleVacancyService.java
@@ -1,5 +1,6 @@
 package org.dreamjob.service;
 
+import org.dreamjob.dto.FileDto;
 import org.dreamjob.model.Vacancy;
 import org.dreamjob.repository.VacancyRepository;
 import org.springframework.stereotype.Service;
@@ -12,23 +13,46 @@ public class SimpleVacancyService implements VacancyService {
 
     private final VacancyRepository vacancyRepository;
 
-    public SimpleVacancyService(VacancyRepository vacancyRepository) {
-        this.vacancyRepository = vacancyRepository;
+    private final FileService fileService;
+
+    public SimpleVacancyService(VacancyRepository sql2oVacancyRepository, FileService fileService) {
+        this.vacancyRepository = sql2oVacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return vacancyRepository.save(vacancy);
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        var file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-        return vacancyRepository.deleteById(id);
+        var fileOptional = findById(id);
+        if (fileOptional.isEmpty()) {
+            return false;
+        }
+        var isDeleted = vacancyRepository.deleteById(id);
+        fileService.deleteById(fileOptional.get().getFileId());
+        return isDeleted;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        var isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+        var oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        var isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
